Got everything working, along with task refactor
