+++ firmware/.cargo/config.toml
@@ -39,9 +39,9 @@ rustflags = [
 
 [build]
 # We compile without atomics, because its faster than using atomic trap handler
-target = "riscv32imc-unknown-none-elf"
+# target = "riscv32imc-unknown-none-elf"
 # target = "xtensa-esp32-none-elf"
-# target = "thumbv7em-none-eabihf"
+target = "thumbv7em-none-eabihf"
 
 [unstable]
 build-std = ["core", "alloc"]
