+++ firmware/src/aliases.rs
@@ -4,6 +4,8 @@ pub mod ඞ {
 	pub use esp32_hal::Delay as DelayConcrete;
 
 	pub type I2cConcrete<'a> = esp32_hal::i2c::I2C<esp32_hal::pac::I2C0>;
+
+	pub type BbqPeripheral<'a> = ();
 }
 
 #[cfg(feature = "mcu-esp32c3")]
@@ -12,6 +14,8 @@ pub mod ඞ {
 	pub use esp32c3_hal::Delay as DelayConcrete;
 
 	pub type I2cConcrete<'a> = esp32c3_hal::i2c::I2C<esp32c3_hal::pac::I2C0>;
+
+	pub type BbqPeripheral<'a> = ();
 }
 
 #[cfg(feature = "mcu-nrf52840")]
@@ -29,6 +33,11 @@ pub mod ඞ {
 		embassy_nrf::peripherals::USBD,
 		embassy_nrf::usb::PowerUsb,
 	>;
+
+	#[cfg(feature = "log-usb-serial")]
+	pub type BbqPeripheralConcrete<'a> = UsbDriverConcrete<'a>;
+	#[cfg(feature = "log-uart")]
+	pub type BbqPeripheralConcrete<'a> = UartConcrete<'a>;
 }
 
 pub trait I2c:
